+++ Gemfile
@@ -4,6 +4,5 @@ source 'https://rubygems.org'
 
 git_source(:github) { |repo_name| "https://github.com/#{repo_name}" }
 
-# gem "rails"
 gem 'fasterer', '~> 0.4.1'
 gem 'rubocop', '~> 0.60.0'
