+++ autoload/autoload.rb
@@ -1,10 +1,10 @@
 require 'yaml'
-require_relative '../Errors/errors_catcher.rb'
+require_relative '../errors/errors_catcher.rb'
 require_relative '../modules/validation.rb'
-require_relative '../controllers/author_controller.rb'
-require_relative '../controllers/book_controller.rb'
-require_relative '../controllers/reader_controller.rb'
-require_relative '../controllers/order_controller.rb'
+require_relative '../entities/author.rb'
+require_relative '../entities/book.rb'
+require_relative '../entities/reader.rb'
+require_relative '../entities/order.rb'
 require_relative '../modules/statistic_module.rb'
 require_relative '../modules/database_module.rb'
-require_relative '../controllers/library_controller.rb'
+require_relative '../entities/library.rb'
