+++ entities/author.rb
