+++ modules/statistic_module.rb
@@ -1,13 +1,20 @@
 module Statistics
   def top_reader(orders, num = 1)
-    orders.group_by(&:reader).sort_by { |_key, value| -value.count }.first(num).to_h.keys
+    counter(num, orders, :reader)
   end
 
   def most_popular_books(orders, num = 1)
-    orders.group_by(&:book).sort_by { |_key, value| -value.count }.first(num).to_h.keys
+    counter(num, orders, :book)
   end
 
   def readers_the_most_popular_books(orders, num = 3)
-    orders.select { |order| most_popular_books(orders, num).include? order.book }.map(&:reader).uniq.count
+    books = most_popular_books(orders, num)
+    orders.select { |order| books.include? order.book }.map(&:reader).uniq.count
+  end
+
+  private
+
+  def counter(num, orders, item)
+    puts orders.group_by(&item).sort_by { |_key, value| -value.count }.first(num).to_h.keys
   end
 end
